Simplify logic by pre-reading body
Reverts the logic where the user determines to read the body or not.
Now the body will always be read (and parsed if chunked). This massively
increases the simplicity of the architecture. This also removes several
footguns where the user may attempt to read the body twice.

As the body is allocated using a stack allocator, it doesn't allocate
on the heap until it reaches the max buffer size, meaning the performance impact
is almost non-existant until the max buffer size (or stack size) is reached.