Add 'form' example and fix multipart/form parsing
Inititally I thought each line in a multipart form body used \r\n for line endings.
However, it behaves like regular bodies and uses \n for line endings (atleast that's consistant).

This means we could potentially simplify the parsing a little bit and make it faster.
However, that's a 'problem' for later.